Move files to where flask-restful recommends
Closes #4.
